requirements.txt
beautifulsoup4
requests
selenium
termcolor
python-dotenv
twilio